Cleaning backend for the new scope
