--- app/helpers/day_of_weeks_helper.rb
@@ -1,2 +1,0 @@
-module DayOfWeeksHelper
-end
