--- app/helpers/schedules_helper.rb
@@ -1,2 +1,0 @@
-module SchedulesHelper
-end
