--- app/serializers/day_of_week_serializer.rb
@@ -1,3 +1,0 @@
-class DayOfWeekSerializer < ActiveModel::Serializer
-  attributes :id, :day, :value, :departure_time, :return_time, :schedule
-end
