+++ app/serializers/ride_serializer.rb
@@ -1,9 +1,7 @@
 class RideSerializer < ActiveModel::Serializer
   attributes :id, :title, :origin, :destination, :route_distance, :route_time, :total_seats,
-  :schedule_ride, :is_finished, :is_subsistence_allowance, :description, :driver,
+  :is_finished, :is_subsistence_allowance, :description, :driver,
   :vehicle, :passengers_name, :passengers_photo, :created_at, :updated_at
-
-  has_many :schedules
 
   def passengers_name
     object.passengers.map do |passenger|
