--- db/migrate/20160527201950_change_ride_table.rb
@@ -1,5 +1,0 @@
-class ChangeRideTable < ActiveRecord::Migration
-  def change
-    rename_column :rides, :destiny, :destination
-  end
-end
