--- db/migrate/20160528034022_change_all_ride_table.rb
@@ -1,7 +1,0 @@
-class ChangeAllRideTable < ActiveRecord::Migration
-  def change
-  	rename_column :rides, :distance, :route_distance
-  	rename_column :rides, :time, :route_time
-  	remove_column :rides, :departure_time, :return_time, :date
-  end
-end
