--- db/migrate/20160528035136_remove_some_columns.rb
@@ -1,6 +1,0 @@
-class RemoveSomeColumns < ActiveRecord::Migration
-  def change
-  	remove_column :rides, :return_time
-  	remove_column :rides, :date
-  end
-end
