--- db/migrate/20160528040504_remove_column_day_of_week.rb
@@ -1,5 +1,0 @@
-class RemoveColumnDayOfWeek < ActiveRecord::Migration
-  def change
-  	remove_column :schedules, :day_of_week
-  end
-end
