--- db/migrate/20160528041909_change_route_time_type.rb
@@ -1,5 +1,0 @@
-class ChangeRouteTimeType < ActiveRecord::Migration
-  def change
-  	change_column :rides, :route_time, :string
-  end
-end
