+++ spec/models/ride_spec.rb
@@ -7,7 +7,7 @@ RSpec.describe Ride, :type => :model do
     @other_ride_attrs = FactoryGirl.attributes_for(:ride, driver: @user.driver)
   end
 
-  it { expect(@ride).to respond_to(:id, :title, :origin, :destination, :route_distance, :route_time, :total_seats, :schedule_ride, :is_finished, :is_subsistence_allowance, :description, :vehicle, :driver, :created_at, :updated_at) }
+  it { expect(@ride).to respond_to(:id, :title, :origin, :destination, :route_distance, :route_time, :total_seats, :is_finished, :is_subsistence_allowance, :description, :vehicle, :driver, :created_at, :updated_at) }
 
   describe "attributes:" do
 
@@ -134,18 +134,6 @@ RSpec.describe Ride, :type => :model do
         subject["route_time"] = nil
         expect(Ride.new(subject)).not_to be_valid
         subject["route_time"] = "20 minutos"
-        expect(Ride.new(subject)).to be_valid
-      end
-    end
-
-    describe "schedule_ride" do
-
-      subject { @other_ride_attrs }
-
-      it "must be given" do
-        subject["schedule_ride"] = nil
-        expect(Ride.new(subject)).not_to be_valid
-        subject["schedule_ride"] = true
         expect(Ride.new(subject)).to be_valid
       end
     end
